Extract CP852 to its own file
