+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/encoding/charset/CP852.kt
@@ -1,0 +1,152 @@
+/*
+ *    Copyright 2024 Ondřej Karmazín
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package cz.multiplatform.escpos4k.core.encoding.charset
+
+/** Page 18 [CP852: Latin 2] */
+public data object CP852 : Charset(18, "IBM852") {
+  override val mapping: Map<Char, Int> =
+      mapOf(
+          '\u00C7' to 0x80,
+          '\u00FC' to 0x81,
+          '\u00E9' to 0x82,
+          '\u00E2' to 0x83,
+          '\u00E4' to 0x84,
+          '\u016F' to 0x85,
+          '\u0107' to 0x86,
+          '\u00E7' to 0x87,
+          '\u0142' to 0x88,
+          '\u00EB' to 0x89,
+          '\u0150' to 0x8A,
+          '\u0151' to 0x8B,
+          '\u00EE' to 0x8C,
+          '\u0179' to 0x8D,
+          '\u00C4' to 0x8E,
+          '\u0106' to 0x8F,
+          '\u00C9' to 0x90,
+          '\u0139' to 0x91,
+          '\u013A' to 0x92,
+          '\u00F4' to 0x93,
+          '\u00F6' to 0x94,
+          '\u013D' to 0x95,
+          '\u013E' to 0x96,
+          '\u015A' to 0x97,
+          '\u015B' to 0x98,
+          '\u00D6' to 0x99,
+          '\u00DC' to 0x9A,
+          '\u0164' to 0x9B,
+          '\u0165' to 0x9C,
+          '\u0141' to 0x9D,
+          '\u00D7' to 0x9E,
+          '\u010D' to 0x9F,
+          '\u00E1' to 0xA0,
+          '\u00ED' to 0xA1,
+          '\u00F3' to 0xA2,
+          '\u00FA' to 0xA3,
+          '\u0104' to 0xA4,
+          '\u0105' to 0xA5,
+          '\u017D' to 0xA6,
+          '\u017E' to 0xA7,
+          '\u0118' to 0xA8,
+          '\u0119' to 0xA9,
+          '\u00AC' to 0xAA,
+          '\u017A' to 0xAB,
+          '\u010C' to 0xAC,
+          '\u015F' to 0xAD,
+          '\u00AB' to 0xAE,
+          '\u00BB' to 0xAF,
+          '\u2591' to 0xB0,
+          '\u2592' to 0xB1,
+          '\u2593' to 0xB2,
+          '\u2502' to 0xB3,
+          '\u2524' to 0xB4,
+          '\u00C1' to 0xB5,
+          '\u00C2' to 0xB6,
+          '\u011A' to 0xB7,
+          '\u015E' to 0xB8,
+          '\u2563' to 0xB9,
+          '\u2551' to 0xBA,
+          '\u2557' to 0xBB,
+          '\u255D' to 0xBC,
+          '\u017B' to 0xBD,
+          '\u017C' to 0xBE,
+          '\u2510' to 0xBF,
+          '\u2514' to 0xC0,
+          '\u2534' to 0xC1,
+          '\u252C' to 0xC2,
+          '\u251C' to 0xC3,
+          '\u2500' to 0xC4,
+          '\u253C' to 0xC5,
+          '\u0102' to 0xC6,
+          '\u0103' to 0xC7,
+          '\u255A' to 0xC8,
+          '\u2554' to 0xC9,
+          '\u2569' to 0xCA,
+          '\u2566' to 0xCB,
+          '\u2560' to 0xCC,
+          '\u2550' to 0xCD,
+          '\u256C' to 0xCE,
+          '\u00A4' to 0xCF,
+          '\u0111' to 0xD0,
+          '\u0110' to 0xD1,
+          '\u010E' to 0xD2,
+          '\u00CB' to 0xD3,
+          '\u010F' to 0xD4,
+          '\u0147' to 0xD5,
+          '\u00CD' to 0xD6,
+          '\u00CE' to 0xD7,
+          '\u011B' to 0xD8,
+          '\u2518' to 0xD9,
+          '\u250C' to 0xDA,
+          '\u2588' to 0xDB,
+          '\u2584' to 0xDC,
+          '\u0162' to 0xDD,
+          '\u016E' to 0xDE,
+          '\u2580' to 0xDF,
+          '\u00D3' to 0xE0,
+          '\u00DF' to 0xE1,
+          '\u00D4' to 0xE2,
+          '\u0143' to 0xE3,
+          '\u0144' to 0xE4,
+          '\u0148' to 0xE5,
+          '\u0160' to 0xE6,
+          '\u0161' to 0xE7,
+          '\u0154' to 0xE8,
+          '\u00DA' to 0xE9,
+          '\u0155' to 0xEA,
+          '\u0170' to 0xEB,
+          '\u00FD' to 0xEC,
+          '\u00DD' to 0xED,
+          '\u0163' to 0xEE,
+          '\u00B4' to 0xEF,
+          '\u00AD' to 0xF0,
+          '\u02DD' to 0xF1,
+          '\u02DB' to 0xF2,
+          '\u02C7' to 0xF3,
+          '\u02D8' to 0xF4,
+          '\u00A7' to 0xF5,
+          '\u00F7' to 0xF6,
+          '\u00B8' to 0xF7,
+          '\u00B0' to 0xF8,
+          '\u00A8' to 0xF9,
+          '\u02D9' to 0xFA,
+          '\u0171' to 0xFB,
+          '\u0158' to 0xFC,
+          '\u0159' to 0xFD,
+          '\u25A0' to 0xFE,
+          '\u00A0' to 0xFF,
+      )
+}
